add new code layout for list_cpu_features
